refactor: giftsRender에서 데이터 패치를 하도록 변경
