+++ src/components/GiftsRender.tsx
@@ -1,15 +1,27 @@
-import type { Gift } from "@/types/gift";
 import GiftsList from "./GiftsList";
 import LoadingSpinner from "@/components/common/LoadingSpinner";
 import BoxMessage from "./common/BoxMessage";
+import { fetchProductsRanking } from "@/api/products";
+import { useCallback } from "react";
+import useApiRequest from "@/hooks/useApiRequest";
+import type { TargetType, RankType } from "@/types/gift";
 
 type GiftsRenderProps = {
-  gifts: Gift[] | undefined;
-  isLoading: boolean;
-  isError: boolean;
+  selectedTypes: {
+    targetType: TargetType;
+    rankType: RankType;
+  };
 };
 
-const GiftsRender = ({ gifts, isLoading, isError }: GiftsRenderProps) => {
+const GiftsRender = ({ selectedTypes }: GiftsRenderProps) => {
+  const requestFn = useCallback(() => {
+    return fetchProductsRanking(
+      selectedTypes.targetType,
+      selectedTypes.rankType,
+    );
+  }, [selectedTypes]);
+  const { data: gifts, isLoading, isError } = useApiRequest({ requestFn });
+
   if (isLoading) {
     return <LoadingSpinner height="266px" />;
   }
